setup.sh: network: try less hard to configure the network
/etc/network/interfaces is legacy as of Ubuntu 17.10, and we should
take advantage of netplan as much as possible. To that end, we should
simplify and make our code safer and easier to reason about.

Specifically, we can't easily guarantee that our sed statement is
going to reconfigure the network interface properly. Since it is the
network interface, I'd rather not get this wrong. For that reason,
I've made it append-only the interface configuration and recommend
manual commands to run.

Tinkerbell-Legacy-Original-SHA1: 8f28f156d89c85b2802e0980eaa72ddfc283c361