Change packet.Hardware to interface instead of pointer-to-interface
By using a regular interface instead, the readability of the code jumps
up quite a bit. Pointer to interface is not needed in this case as we
aren't modifying the pointer as a return value (like json.Decode and co
do).

Tinkerbell-Legacy-Original-SHA1: 093fd00af1a44e9f8033ddb5e3e2e6ad862feef8