Replace go fmt with goimports
goimports is a superset of go fmt, it also ensures the imports are
grouped and sorted nicely. Most importantly it has a `-d`/diff mode that
we can use to test for changes more easily than what would be needed if
we stayed with `go fmt`.

Signed-off-by: Manuel Mendez <manny@packet.com>

Tinkerbell-Legacy-Original-SHA1: ce89835a5b3808d4797064d0131660cc89d44f5c