Change TestGetMetadataTinkerbell from packet.Metadata to map[string]interface{}
When the default handling for /metadata changes in an upcoming commit
unmarshalling the json to packet.Metadata is not going to work. So we
change the unmarshalling destination here instead so the next commit
is doing only one thing.

Signed-off-by: Manuel Mendez <mmendez@equinix.com>

Tinkerbell-Legacy-Original-SHA1: a678e910b803c426c5ff95f19a06b807ec603455