Dockerfile: Only COPY binary for TARGET
No need to copy all the binaries only to keep one of them. This was also
already buggy as `uname -m` was showing x86_64 for both the x86_64 and
386 builds.

This is making use of the buildx multi-platform provided ARGS, a normal
`docker build` user would have to pass these values in manually. This
seems like a step backwards compared to `uname -m` but as I mentioned
previously, that was already broken.

Signed-off-by: Manuel Mendez <mmendez@equinix.com>

Tinkerbell-Legacy-Original-SHA1: b40ecfe95261101944fa5aa7c3573abb7cbb4f3f