Makefile: small bits of refactors
Change variables to be defined with `:=` so they are "Simply expanded
variables" and not recursive... which basically just means they are
evaluated immediately instead of when used. This is the least surprising
behavior.

While here I changed the names of some variables so its shorter and a
little clearer.

I also dropped use of GOOS as it doesn't seem used and moved CGO_ENABLED
out of the recipe since it'll be re-used soon.

Signed-off-by: Manuel Mendez <mmendez@equinix.com>

Tinkerbell-Legacy-Original-SHA1: cc9831c09aa567929e88580cb2ebaaf960d29157