dhcp: Minimize setPXEFilename logic / special cases
This patch changes setPXEFilename so that it mostly only cares about AllowPXE
and the machine type, but not exclusively just those two bits of info.

We still care about an instance being active or not because we don't want to
set any filename if the machine has an instance and is not active. We don't
want to set a filename so that the nic hangs on boot since this is scenario
indicates a machine that has no business even being on, much less trying to
PXE boot. In contrast when AllowPXE == false and instance.state == "active"
and the machine is attempting to PXE boot indicates that the BIOS/FW is setup
with netboot as a boot option prior to booting from disk. In this case we
send a non existent file so that we can short circuit the nic boot attempt
and go on to the next boot option.

Signed-off-by: Manuel Mendez <mmendez@equinix.com>

Tinkerbell-Legacy-Original-SHA1: 9792acfe5ba5dcb7148cf1e4558bff42eb79369b