job: Change setPXEFilename -> getPXEFilename
This way its easier to test since it is now a pure fucntion. I was getting a
panic when trying to rework the not-our-ipxe case for an unknown machine
because it would try to call j.Error(). This is just better code.

Signed-off-by: Manuel Mendez <mmendez@equinix.com>

Tinkerbell-Legacy-Original-SHA1: 42c0297b5d8d3f44b229cdb822b1c3cb8e7b9205