Build IPXE on darwin easier:
Currently building IPXE on a Mac is difficult and
almost entirely manual, if not impossible even with
the nix shell.nix. This will leverage docker to make
building IPXE on Mac behave the same as on Linux with
just the make bindata target.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>

Tinkerbell-Legacy-Original-SHA1: 1725100ace79d1ff559c9ec311f70c179bf8bc28