add context.Context support
This is a first pass of adding context.Context propagation to most of
the relevant code. With this in place, context will now be propagated
down to remote calls to other systems, so traces can show all the
hops through e.g. cacher, tinkerbell, and API.

In addition this enables request cancelation to work as expected by
passing request context to e.g. http requests to API. I only plumbed the
context through to the clients, and do not know if further work is
necessary to get cancelation working. It should be a lot less tedious
now :)

Signed-off-by: Amy Tobey <atobey@equinix.com>

Tinkerbell-Legacy-Original-SHA1: 4f17a5e7bcf498eb355474aa7f47fefa0d607336