add ctx to setPXEFilename and have it add traceparent
The easiest way to send traceparent from DHCP is in the boot_file option
and a nice side-effect is it's the only effective way to send a
traceparent through TFTP. Adding this makes it so the tftp client on the
other side automatically propagates the traceparent to the tftp server.

DANGER: unknown if this will break some PXE clients, might revert after
some testing.

Signed-off-by: Amy Tobey <atobey@equinix.com>

Tinkerbell-Legacy-Original-SHA1: bb25cd3eb840ac8f94f4de7f937a3375cca486e3