Add ability to enable single port mode for the TFTP server:
A standard TFTP server implementation receives requests on port 69 and
allocates a new high port (over 1024) dedicated to that request. In single
port mode, the same port is used for transmit and receive. If the server
is started on port 69, all communication will be done on port 69.
This option is required when running in a container, that doesn't bind to the hosts
network, because this type of dynamic port allocation is not generally supported.

This option is specific to github.com/pin/tftp. The pin/tftp library says this option is
experimental and "Enabling this will negatively impact performance". Please take this into
consideration when using this option.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>

Tinkerbell-Legacy-Original-SHA1: fa3aa698fbb9013a4b355d27343ccf70134f31b7