Stop serving and fetching of grpc cert via /cert path
This was only ever a hack that was moderately ok when it was done in Cacher
when running in an old EM orchestration stack. IT was ok because we *always*
fetched the cert via a "normal" tls connection via a tls terminating proxy.
This was not guaranteed in tinkerbell land and actually exists insecurely in
the sandbox at the moment (see sandbox fetching the cert via http).

Having this here made self-signed certs "easier" but also less safe while making
the normal and desired behavior more difficult. The normal being tinkerbell
using cert generated by a trusted CA. This way certificate revokation can
actually happen, where as with the /cert method all clients would have to
redownload the /cert after figuring out the cert is no longer valid (somehow).

This should never have made it into tinkerbell from cacher.

Signed-off-by: Manuel Mendez <mmendez@equinix.com>

Tinkerbell-Legacy-Original-SHA1: 666425137361c0e5911985bc672dd224e49fefb3