Generalize the hardware client interface
The hardware client is used to interact with a Cacher service and the
Tink service. It embodies gRPC specific types and an ambiguous All()
call that are difficult to model for a Kubernetes data provider.

This change alters the interface to model more generic behavior and
removes gRPC specific types paving the path for implementing a
Kubernetes data provider client.

Signed-off-by: Chris Doherty <chris.doherty4@gmail.com>

Tinkerbell-Legacy-Original-SHA1: 06a45192fb331794dd52d22465dbe5d27c4e88f0