kube: Rework k8s release manifests
We shouldn't be patching a git tracked file as part of the build because
then we'd have to go back and checkout the changes. Unfortunately this means
we need to copy the whole directory tree so that kustomize will pick up the
edited files.

I also fixed a typo that used the TINK_SERVER_IMAGE_NAME for the manager.

I also cleaned up the recursive uses of make since they are considered
harmful and dropped `release` as that should be for the whole project not
just the k8s manifests.

Signed-off-by: Manuel Mendez <github@i.m.mmlb.dev>

Tinkerbell-Legacy-Original-SHA1: d303dc2918df11a526868ff528820d43aeb46d6b