Move GetDiscoveryFromEM from client/packet to client/cacher
Its only meant to be used in cacher mode and not necessarily for anything
to do with packet. Besides that trying to keep it in packet and make use of
it in cacher causes a cyclic import so something needed to be done anyway.

I fiddled around with a few other options like type asserting on
packet.Reporter or asserting on a GetDiscoveryFromEMer interface but both
seemed less useful outside of cacher's needs. I opted to bring this in to
cacher anyway since its only used by cacher and returns a DiscoveryCacher
and just feels like it should be private in cacher.

Changed it to use a `poster` since we actually don't need a reporter.

Signed-off-by: Manuel Mendez <github@i.m.mmlb.dev>

Tinkerbell-Legacy-Original-SHA1: 628f3b7a147d2f65ded012050901dc97634e39eb