cacher: Rename {ZeroMAC,OnesMAC} -> {MinMAC,MaxMAC}
OnesMAC was a confusing name because it made me think it was
"11:11:11:11:11:11" not all ones as in binary. Its important to be the MAX mac
(chose to name similar to math.MaxInt) because we're interested in finding
the lowest mac as the fallback option.

Renamed ZeroMAC so it matches up with naming in package math too.

Signed-off-by: Manuel Mendez <github@i.m.mmlb.dev>

Tinkerbell-Legacy-Original-SHA1: 5e7645fa9833be36d3602296a36c85c13f20970c