Use string type for VLANID:
This allows a default value of "" to be used
instead of 0 when the VLANID type is int. As 0
is a valid vlan id, we dont want that to be the default
as it would make it impossible for the VLANID to be "unset"
and therefore not added to DHCP requests, etc.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>

Tinkerbell-Legacy-Original-SHA1: 8effbb87d1325e7e934771907878abdb5dc2314b