Refacfor handlers with frontend's and backends
To create a cleaner separation between frontends and backends this
refactor looks to introduce a set of packages that have a clear scope
and purpose. In doing so we create far more testable code that is
clear in its intent and handling of data between layers.

Signed-off-by: Chris Doherty <chris.doherty4@gmail.com>

Tinkerbell-Legacy-Original-SHA1: 8037e8018a736686970502e2b3428b9d43d4379b