Use controller-runtime's cluster instead of manager
Hegel originally integrated with Kubernetes by using a tink repo defined
client that handled the client construction for a service. The
implementation leveraged the controller-runtime manager data structure
that comes with a lot more than whats needed by a service. Consequently,
it needs to mute features such as metrics endpoints etc.

This change switches to the cluster object which is much lighter weight
and focus' on representing a cluster (with caching of obects) instead of
a fully fledged controller manager.

Signed-off-by: Chris Doherty <chris.doherty4@gmail.com>

Tinkerbell-Legacy-Original-SHA1: 4bda0aa80f6715a219f6babaed1717456f67a911