Remove otel, use a lightweight journal:
Otel was working well, but it was heavy and
required a otel specific log collector.
The journal is a lot lighter weight. Journal
entries are collected during a reconcile and
logged once at the end. As this produces a good
sized log line per reconcile, debug logging was added
so that this is opt in only.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>

Tinkerbell-Legacy-Original-SHA1: 06fc9f5e70e54413d93501ebad581cbaecff66d9