Improve rbac setting by using Helm globals:
This allows all rbac to be toggled between
Role and ClusterRole. It also simplifies the
trustedProxies for Hegel and Smee and the
loadBalancerIP/publicIP shared between the stack
and Smee. The Helm install experience is improved,
in my opinion.

I also removed all the required fields in the
values.schema.json so that flexibility of configuration
is not limited.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>

Tinkerbell-Legacy-Original-SHA1: ebac3403baacedcf1ad4c881aa861f1c040390ba