Default to using local persistent volume:
This allows the tink-stack deployment to reference only a PVC.
If a user wants a production grade experience they
should provide a PVC backed by a production grade PV.

What used to be a Kubernetes Job to download HookOS, is
now a container in the tink-stack pod. This removes the
challenges of getting a Kubernetes Job to download HookOS
on the same node as the tink-stack pod. This is only an issue
with the local persistent volume.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>

Tinkerbell-Legacy-Original-SHA1: e95a0375f4acf4f94935bdd79913ff71a7a5290b