Memory safety update:
If the request is a partial content request,
we need to validate the Content-Range header.
Because we read the entire response body into
memory for patching, we need to ensure that the
Content-Range is within a reasonable size.
For now, we are limiting the size to 500Kb.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>

Tinkerbell-Legacy-Original-SHA1: c6e17f70318099de56bda0e33cb39402cac95d34