Make default be the default namespace for leader election:
Rufio and Tink controller need to create lease objects
for leader election. When Tinkerbell is run outside of
a Kubernetes cluster this namespace must be explicitly
defined. When Tinkerbell is run in cluster, if a namespace
is not define, the namespace where Tinkerbell is running
is used.

Also, in the logging output, trim the source file name
for 3rd party packages.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>