Fix Agent logic:
There were a few places that weren't behaving properly
and caused that Workflows that had failed, could not be
delete and rerun from the same Agent.

Also, there was a backwards incompatible change to
parsing the `command` from the Worflow Action. The
first value in the command was being used to overwrite
the entrypoint in the container. This caused Actions
to fail that normally would not.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>