Package reorg:
Simpler top level improves the cognitive load
for understanding and contributing to the code base.

Added a Makefile goal/target to produce a dependency
graph of the Tinkerbell and Agent binaries. This is
a first step aimed helping ensure we understand our
dependencies graph and don't add additional
spaghetti code.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>