Improve log handling for embedded services:
Kube-apiserver and etcd now have more standardized
logging output with the other services. kube-apiserver
lets us pass in a logr.Logger so it's the exact same logger
as the other services. etcd does not use logr.Logger. etcd
only uses zap. It does allow us to customize it. So a best
effort to get it to look like the other logger is put forth.

Some linting fixes have been done and the order of Makefile
variables was fixed so that dependent variables are
instantiated first.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>