Use PreferDriver instead of PreferProtocol:
The API spec defines PreferredOrder to be the driver
name not the protocol name. This causes the "gofish"
provider to not change the order as that is the driver
name not the protocol. The protocol is "redfish".
Changing the implementation to use PreferDriver is
chosen so as not to have to change the API spec validation.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>

Tinkerbell-Legacy-Original-SHA1: 9d0406e3140535082d00212239c1cdf1e5788759